refactor: Move parts specific for workspace folder from `MetalsLspServer`
