+++ metals/src/main/scala/scala/meta/internal/metals/FallbackMetalsLspService.scala
@@ -8,11 +8,16 @@ import scala.concurrent.ExecutionContextExecutorService
 import scala.concurrent.Future
 
 import scala.meta.internal.builds.ShellRunner
+import scala.meta.internal.metals.Indexer.BuildTool
 import scala.meta.internal.metals.MetalsEnrichments._
 import scala.meta.internal.metals.clients.language.ConfiguredLanguageClient
 import scala.meta.internal.metals.doctor.HeadDoctor
+import scala.meta.internal.metals.watcher.FileWatcher
+import scala.meta.internal.metals.watcher.NoopFileWatcher
+import scala.meta.internal.mtags.Semanticdbs
 import scala.meta.io.AbsolutePath
 
+import ch.epfl.scala.bsp4j.DidChangeBuildTarget
 import org.eclipse.lsp4j.DidCloseTextDocumentParams
 import org.eclipse.lsp4j.InitializeParams
 
@@ -27,7 +32,6 @@ class FallbackMetalsLspService(
     focusedDocument: () => Option[AbsolutePath],
     shellRunner: ShellRunner,
     timerProvider: TimerProvider,
-    initTreeView: () => Unit,
     folder: AbsolutePath,
     folderVisibleName: Option[String],
     headDoctor: HeadDoctor,
@@ -44,7 +48,6 @@ class FallbackMetalsLspService(
       focusedDocument,
       shellRunner,
       timerProvider,
-      initTreeView,
       folder,
       folderVisibleName,
       headDoctor,
@@ -56,18 +59,14 @@ class FallbackMetalsLspService(
   buildServerPromise.success(())
   indexingPromise.success(())
 
+  val pauseables: Pauseable = Pauseable.fromPausables(
+    parseTrees ::
+      compilations.pauseables
+  )
+
   private val files: AtomicReference[Set[AbsolutePath]] = new AtomicReference(
     Set.empty
   )
-
-  override def maybeImportScript(path: AbsolutePath): Option[Future[Unit]] = {
-    if (!path.isScala) None
-    else {
-      val prev = files.getAndUpdate(_ + path)
-      if (prev.contains(path)) None
-      else Some(scalaCli.start(path))
-    }
-  }
 
   override def didClose(params: DidCloseTextDocumentParams): Unit = {
     val path = params.getTextDocument.getUri.toAbsolutePath
@@ -75,6 +74,44 @@ class FallbackMetalsLspService(
     super.didClose(params)
     scalaCli.stop(path)
   }
+
+  protected def buildData(): Seq[BuildTool] =
+    scalaCli.lastImportedBuilds.map {
+      case (lastImportedBuild, buildTargetsData) =>
+        Indexer
+          .BuildTool("scala-cli", buildTargetsData, lastImportedBuild)
+    }
+  def buildHasErrors(path: AbsolutePath): Boolean = false
+  protected def check(): Unit = {}
+  protected def didCompileTarget(
+      report: ch.epfl.scala.bsp4j.CompileReport
+  ): Unit = {}
+  def fileWatcher: FileWatcher = NoopFileWatcher
+  def getTargetsInfoForReports(): List[Map[String, String]] = Nil
+  def maybeFixAndLoad(load: () => Future[Unit]): Future[Unit] =
+    for {
+      _ <-
+        if (!path.isScala) Future.unit
+        else {
+          val prev = files.getAndUpdate(_ + path)
+          if (prev.contains(path)) Future.unit
+          else scalaCli.start(path)
+        }
+      _ <- load()
+    } yield ()
+  protected def onBuildTargetChanges(params: DidChangeBuildTarget): Unit = {
+    compilations.cancel()
+    val scalaCliAffectedServers = params.getChanges.asScala
+      .flatMap { change =>
+        buildTargets.buildServerOf(change.getTarget)
+      }
+      .flatMap(conn => scalaCli.servers.find(_ == conn))
+    importAfterScalaCliChanges(scalaCliAffectedServers)
+  }
+  protected def onInitialized(): Future[Unit] = Future.unit
+  def onMissingSemanticDB(path: AbsolutePath): Unit = {}
+  def scalaCliDirOrFile(path: AbsolutePath): AbsolutePath = path
+  protected val semanticdbs: Semanticdbs = interactiveSemanticdbs
 
 }
