+++ metals/src/main/scala/scala/meta/internal/metals/testProvider/TestSuitesProvider.scala
@@ -51,7 +51,7 @@ final class TestSuitesProvider(
     buildTargetClasses: BuildTargetClasses,
     trees: Trees,
     symbolIndex: GlobalSymbolIndex,
-    semanticdbs: Semanticdbs,
+    semanticdbs: () => Semanticdbs,
     buffers: Buffers,
     clientConfig: ClientConfiguration,
     userConfig: () => UserConfiguration,
@@ -65,11 +65,11 @@ final class TestSuitesProvider(
   private val index = new TestSuitesIndex
   private val junitTestFinder = new JunitTestFinder
   private val munitTestFinder =
-    new MunitTestFinder(trees, symbolIndex, semanticdbs)
+    new MunitTestFinder(trees, symbolIndex, semanticdbs())
   private val scalatestTestFinder =
-    new ScalatestTestFinder(trees, symbolIndex, semanticdbs)
+    new ScalatestTestFinder(trees, symbolIndex, semanticdbs())
   private val weaverCatsEffect =
-    new WeaverCatsEffectTestFinder(trees, symbolIndex, semanticdbs)
+    new WeaverCatsEffectTestFinder(trees, symbolIndex, semanticdbs())
 
   private def isExplorerEnabled = clientConfig.isTestExplorerProvider() &&
     userConfig().testUserInterface == TestUserInterfaceKind.TestExplorer
@@ -302,7 +302,7 @@ final class TestSuitesProvider(
       doc: Option[TextDocument],
   ): Seq[AddTestCases] = {
     doc
-      .orElse(semanticdbs.textDocument(path).documentIncludingStale)
+      .orElse(semanticdbs().textDocument(path).documentIncludingStale)
       .map { semanticdb =>
         suites.flatMap { suite =>
           val testCases = suite.framework match {
